refactor: rename folder to multitype queue
