+++ src/agent/multitype_queue/tests/CMakeLists.txt
@@ -1,17 +1,17 @@
 find_package(GTest REQUIRED)
 
-add_executable(test_AgentQueue queue_test.cpp)
-target_link_libraries(test_AgentQueue PUBLIC
-    AgentQueue
+add_executable(test_MultiTypeQueue multitype_queue_test.cpp)
+target_link_libraries(test_MultiTypeQueue PUBLIC
+    MultiTypeQueue
     GTest::gtest
     GTest::gtest_main
     GTest::gmock
     GTest::gmock_main)
-add_test(NAME AgentQueueTest COMMAND test_AgentQueue)
+add_test(NAME AgentQueueTest COMMAND test_MultiTypeQueue)
 
 add_executable(test_sqlitestorage sqlitestorage_test.cpp)
 target_link_libraries(test_sqlitestorage
-    AgentQueue
+    MultiTypeQueue
     SQLiteCpp
     GTest::gtest
     GTest::gtest_main
