+++ src/agent/multitype_queue/tests/multitype_queue_test.cpp
@@ -12,8 +12,8 @@
 #include <boost/asio.hpp>
 #include <boost/asio/experimental/co_spawn.hpp>
 
-#include "agent_queue.hpp"
-#include "queue_test.hpp"
+#include "multitype_queue.hpp"
+#include "multitype_queue_test.hpp"
 
 using json = nlohmann::json;
